pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
